backend/requirements.txt
# =========================
# Web API
# =========================
fastapi==0.115.8
uvicorn==0.34.0
starlette==0.45.3
pydantic==2.12.4
pydantic-settings==2.10.1
python-dotenv==1.1.0
orjson==3.10.13
gunicorn==23.0.0

# =========================
# Data & ML Core
# =========================
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.6.0
pandas==2.2.3
joblib==1.4.2

# =========================
# NLP & Embeddings
# =========================
torch==2.5.1
transformers==4.49.0
sentence-transformers==5.1.2
tokenizers==0.21.0
safetensors==0.5.2
huggingface-hub==0.28.1

# =========================
# TensorFlow (LEGACY / SAFE)
# =========================
tensorflow
tf-keras==2.19.0
h5py==3.13.0
protobuf==5.29.1

# =========================
# Vector Search
# =========================
faiss-cpu==1.12.0

# =========================
# HTTP Client
# =========================
requests==2.32.3
httpx==0.28.1
